Add FeedbackDAO,Controller and jsp

+++ src/main/java/team5/dao/UserDao/UserDao.java
@@ -69,7 +69,7 @@ public class UserDao implements EntityDao<User>{
         return users;
     }
 
-    public long getCount(){
+    public int getCount(){
         return findAll().size();
     }
